ata-tidy-fixes-01(07) -bugprone-implicit-widening-of-multiplication-result, part 2

+++ src/solver/conv_asm_implicit_gemm_gtc_bwd_nhwc.cpp
@@ -436,7 +436,7 @@ GetImplicitGemmGtcDynamicBwdXdlopsNHWCKernel(
     const auto gemm_n = c / group;
 
     size_t block_size = config.BlockSize();
-    size_t grid_size  = group * integer_divide_ceil(gemm_m, config.gemm_m_per_block) *
+    size_t grid_size  = static_cast<size_t>(group) * integer_divide_ceil(gemm_m, config.gemm_m_per_block) *
                        integer_divide_ceil(gemm_n, config.gemm_n_per_block) *
                        (1 << config.gemm_k_global_split);
     if(config.multihead != 0)
